emergency touch fixes
iphone was "suddenly" broken -- a viewport setting to prevent zooming
was fixing it at crazy-huge scale. As I fixed it I noticed problems with
fixed-scroll keys. iPhone was setting the height too short in landscape,
and wasn't scrolling it! (once I got it to scroll, the rest of the time
it didn't work)